timer finally working and integrated
