+++ components/Timer.js
@@ -1,5 +1,6 @@
-import { ButtonContainer, SmallerHeadline, StandardButton } from "@/styledcomponents"
-import { useState, useEffect } from "react";
+import { SmallerHeadline } from "@/styledcomponents"
+import { formatDuration } from "@/utils";
+import { useState, useEffect, useCallback } from "react";
 import styled from "styled-components"
 
 const TimerSection = styled.section`
@@ -7,96 +8,58 @@ width: 8rem;
 `;
 
 const TimeDisplay = styled.div`
-text-align: right; 
-font-size: 1rem;
-font-weight: 500; 
-margin: 0 0 0 1rem;
-padding: .5rem;
-
+    text-align: right; 
+    font-size: 1rem;
+    font-weight: 500; 
+    padding: .5rem;
+    margin: .5rem 0 0 0;
+    height: 2rem;
+    border-radius: 4px;
+    border: 1px solid darkorange;
+    background-color: white;
 `;
-export default function Timer({runTimer, resetTimer}) {
+export default function Timer({runTimer, resetTimer, sendTime}) {
     const [timeTools, setTimeTools] = useState({ start: 0, running: false })
     const { start, running,  interval1 } = timeTools;
     const [timespan, setTimespan] = useState(0);
-    const [hundredth, setHundredth] = useState(0);
-
 
     function advancedTiming(run) {
-         if (run === true) {
-            const first = Date.now();
-            setTimeTools({ start: first });
+        if (run === true) {
+           const first = Date.now();
+           setTimeTools({ start: first });
 
-            function updateTimespan() {
-                const newTimespan = (Date.now() - first);
-                // const roundedTime = Math.round(newTimespan);
-                setTimespan(newTimespan);
-
-                function fakeHundredth() {
-                    const numbers = [...Array(100).keys()];
-                   numbers.forEach((number) => setTimeout(setHundredth, 10 * number, number));        
-                }
-               fakeHundredth();
-            }
-          
-            const newIntervalId = setInterval(updateTimespan, 1000);
-            setTimeTools({ ...timeTools, interval1: newIntervalId })
+           function updateTimespan() {
+               const newTimespan = (Date.now() - first);
+               setTimespan(newTimespan);
+           }
          
-        } else { 
-            setTimeTools({ ...timeTools, running: false });
-            const newIntervalId = interval1;
-             clearInterval(newIntervalId);
-            //  const finalTime = (Date.now() - start);
-             const finalTime = formatDuration(timespan, true);
-             const finalHundredth = finalTime.slice(-2)
-             setTimeout(()=>setHundredth(finalHundredth), 99);
-             console.log(finalTime);
-             console.log(finalTime.slice(-2));
-        }
-      
-    }
-
-    useEffect(()=>{
+           const newIntervalId = setInterval(updateTimespan, 10);
+           setTimeTools({ ...timeTools, interval1: newIntervalId })
+        
+       } else { 
+           setTimeTools({ ...timeTools, running: false });
+           const newIntervalId = interval1;
+            clearInterval(newIntervalId);
+            sendTime(timespan);
+       }
+   }
+  
+    useEffect(() => {
         advancedTiming(runTimer ? true : false);
-        const finalTime = formatDuration(timespan, true);
-        !runTimer && setHundredth(finalTime.slice(-2));
-        !runTimer && console.log(finalTime);
     }, [runTimer])
     
     function resetToZero() {
         setTimespan(0);
-        setHundredth(0);
     }
+
 
     resetTimer && resetToZero();
 
-    function formatTo2digits(number) {
-        const formattedNumber = number.toLocaleString('en-US', {
-            minimumIntegerDigits: 2,
-            useGrouping: false
-                    })
-        return formattedNumber;
-    }
-
-    function formatDuration(number, highRes) {
-        const allSeconds = Math.floor(number / 1000);
-        const minutes = formatTo2digits(Math.floor(allSeconds / 60));
-        const seconds = formatTo2digits(Math.floor(allSeconds % 60));
-        const realHundredth = formatTo2digits(Math.round((number % 1000)/10));
-        const minutesAndSeconds = `${minutes}:${seconds}`;
-        const formattedDuration = highRes? `${minutesAndSeconds},${realHundredth}`: minutesAndSeconds;
-        return formattedDuration;
-}
-   
     return(
         <TimerSection>
-            <SmallerHeadline> Timer</SmallerHeadline>
-            <TimeDisplay> {formatDuration(timespan)},{formatTo2digits(hundredth)}
+            <SmallerHeadline> Time</SmallerHeadline>
+            <TimeDisplay> {formatDuration(timespan, true)}
             </TimeDisplay> 
-            <ButtonContainer>
-            <StandardButton onClick={()=>advancedTiming(true)}>Start</StandardButton>
-            <StandardButton onClick={()=>advancedTiming(false)}>Stop</StandardButton>
-            <StandardButton onClick={resetToZero}>Reset</StandardButton>
-            </ButtonContainer>
         </TimerSection>
     )
 }
